merging additional branch to master

+++ Agar.io/Agar.IO.Client.Console/Handlers/Handler.cs
@@ -1,14 +1,11 @@
 ﻿using Agar.IO.Client.WinForms.Models;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Agar.IO.Client.WinForms.Handlers
 {
     abstract class Handler
     {
-        public float X { get; set; }
-        public float Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
         protected Handler(Position pos)
         {
